[api] refactoring ast
