+++ spl/api/codegen/__init__.py
@@ -1,3 +1,5 @@
 from spl.api.codegen import ast
+from spl.api.codegen import basic
+from spl.api.codegen import evaluation
 from spl.api.codegen import printing
 from spl.api.codegen import utilities
