+++ spl/api/codegen/ast.py
@@ -3,8 +3,6 @@
 
 from collections import OrderedDict
 from itertools import groupby
-import string
-import random
 import numpy as np
 
 from sympy import Basic
@@ -39,8 +37,6 @@ from pyccel.ast import Comment, NewLine
 from pyccel.ast.core      import _atomic
 from pyccel.ast.utilities import build_types_decorator
 
-from .utilities import build_pythran_types_header, variables
-
 
 from sympde.core import Cross_3d
 from sympde.core import Constant
@@ -70,195 +66,23 @@ from sympde.expr import BilinearForm, LinearForm, Functional, BasicForm
 from spl.fem.splines import SplineSpace
 from spl.fem.tensor  import TensorFemSpace
 from spl.fem.vector  import ProductFemSpace
+
+from .basic import SplBasic
+from .evaluation import EvalMapping, EvalField, EvalVectorField
 from .printing import pycode  # TODO remove from here
+from .utilities import random_string
+from .utilities import build_pythran_types_header, variables
+from .utilities import compute_normal_vector, compute_tangent_vector
+from .utilities import select_loops, filter_product
+from .utilities import rationalize_eval_mapping
+from .utilities import compute_atoms_expr_mapping
+from .utilities import compute_atoms_expr_vector_field
+from .utilities import compute_atoms_expr_field
+from .utilities import compute_atoms_expr
+from .utilities import is_vector_field, is_field
 
 
 FunctionalForms = (BilinearForm, LinearForm, Functional)
-
-#==============================================================================
-def random_string( n ):
-    chars    = string.ascii_lowercase + string.digits
-    selector = random.SystemRandom()
-    return ''.join( selector.choice( chars ) for _ in range( n ) )
-
-
-#==============================================================================
-def compute_normal_vector(vector, discrete_boundary, mapping):
-    dim = len(vector)
-    pdim = dim - len(discrete_boundary)
-    if len(discrete_boundary) > 1: raise NotImplementedError('TODO')
-
-    face = discrete_boundary[0]
-    axis = face[0] ; ext = face[1]
-
-    map_stmts = []
-    body = []
-
-    if not mapping:
-
-        values = np.zeros(dim)
-        values[axis] = ext
-
-    else:
-        M = mapping
-        inv_jac = Symbol('inv_jac')
-        det_jac = Symbol('det_jac')
-
-        # ... construct jacobian on manifold
-        lines = []
-        n_row,n_col = M.jacobian.shape
-        range_row = [i for i in range(0,n_row) if not(i == axis)]
-        range_col = range(0,n_col)
-        for i_row in range_row:
-            line = []
-            for i_col in range_col:
-                line.append(M.jacobian[i_col, i_row])
-
-            lines.append(line)
-
-        J = Matrix(lines)
-        # ...
-
-        J = SymbolicExpr(J)
-
-        if dim == 1:
-            raise NotImplementedError('TODO')
-
-        elif dim == 2:
-            J = J[0,:]
-            # TODO shall we use sympy_sqrt here? is there any difference in
-            # Fortran between sqrt and Pow(, 1/2)?
-            j = (sum(J[i]**2 for i in range(0, dim)))**(1/2)
-
-            values = [inv_jac*J[1], -inv_jac*J[0]]
-
-        elif dim == 3:
-
-            x_s = J[0,:]
-            x_t = J[1,:]
-
-            values = Cross_3d(x_s, x_t)
-            j = (sum(J[i]**2 for i in range(0, dim)))**(1/2)
-            values = [inv_jac*v for v in values]
-
-
-        # change the orientation
-        values = [ext*i for i in values]
-
-        map_stmts += [Assign(det_jac, j)]
-        map_stmts += [Assign(inv_jac, 1./j)]
-
-    for i in range(0, dim):
-        body += [Assign(vector[i], values[i])]
-
-#    print(map_stmts)
-#    print(body)
-#    import sys; sys.exit(0)
-
-    return map_stmts, body
-
-#==============================================================================
-def compute_tangent_vector(vector, discrete_boundary, mapping):
-    raise NotImplementedError('TODO')
-
-
-#==============================================================================
-# TODO remove it later
-def filter_loops(indices, ranges, body, discrete_boundary, boundary_basis=False):
-
-    quad_mask = []
-    quad_ext = []
-    if discrete_boundary:
-        # TODO improve using namedtuple or a specific class ? to avoid the 0 index
-        #      => make it easier to understand
-        quad_mask = [i[0] for i in discrete_boundary]
-        quad_ext  = [i[1] for i in discrete_boundary]
-
-        # discrete_boundary gives the perpendicular indices, then we need to
-        # remove them from directions
-
-    dim = len(indices)
-    for i in range(dim-1,-1,-1):
-        rx = ranges[i]
-        x = indices[i]
-        start = rx.start
-        end   = rx.stop
-
-        if i in quad_mask:
-            i_index = quad_mask.index(i)
-            ext = quad_ext[i_index]
-            if ext == -1:
-                end = start + 1
-
-            elif ext == 1:
-                start = end - 1
-            else:
-                raise ValueError('> Wrong value for ext. It should be -1 or 1')
-
-        rx = Range(start, end)
-        body = [For(x, rx, body)]
-    body = fusion_loops(body)
-    return body
-
-
-#==============================================================================
-def select_loops(indices, ranges, body, discrete_boundary, boundary_basis=False):
-
-    quad_mask = []
-    quad_ext = []
-    if discrete_boundary:
-        # TODO improve using namedtuple or a specific class ? to avoid the 0 index
-        #      => make it easier to understand
-        quad_mask = [i[0] for i in discrete_boundary]
-        quad_ext  = [i[1] for i in discrete_boundary]
-
-        # discrete_boundary gives the perpendicular indices, then we need to
-        # remove them from directions
-
-    dim = len(indices)
-    dims = [i for i in range(dim-1,-1,-1) if not( i in quad_mask )]
-
-    for i in dims:
-        rx = ranges[i]
-        x = indices[i]
-        start = rx.start
-        end   = rx.stop
-
-        rx = Range(start, end)
-        body = [For(x, rx, body)]
-
-    body = fusion_loops(body)
-    return body
-
-def fusion_loops(loops):
-    ranges = []
-    indices = []
-    loops_cp = loops
-
-    while len(loops) == 1 and isinstance(loops[0], For):
-
-        loops = loops[0]
-        target = loops.target
-        iterable = loops.iterable
-
-        if isinstance(iterable, Product):
-            ranges  += list(iterable.elements)
-            indices += list(target)
-            if not isinstance(target,(tuple,list,Tuple)):
-                raise ValueError('target must be a list or a tuple of indices')
-
-        elif isinstance(iterable, Range):
-            ranges.append(iterable)
-            indices.append(target)
-        else:
-            raise TypeError('only range an product are supported')
-
-        loops = loops.body
-
-    if len(ranges)>1:
-        return [For(indices, Product(*ranges), loops)]
-    else:
-        return loops_cp
 
 #==============================================================================
 def init_loop_quadrature(indices, ranges, discrete_boundary):
@@ -376,356 +200,6 @@ def init_loop_support(indices_elm, n_elements,
     return stmts
 
 #==============================================================================
-def filter_product(indices, args, discrete_boundary):
-
-    mask = []
-    ext = []
-    if discrete_boundary:
-        # TODO improve using namedtuple or a specific class ? to avoid the 0 index
-        #      => make it easier to understand
-        mask = [i[0] for i in discrete_boundary]
-        ext  = [i[1] for i in discrete_boundary]
-
-        # discrete_boundary gives the perpendicular indices, then we need to
-        # remove them from directions
-
-    dim = len(indices)
-    args = [args[i][indices[i]] for i in range(dim) if not(i in mask)]
-
-    return Mul(*args)
-
-
-#==============================================================================
-def compute_atoms_expr(atom, indices_quad, indices_test,
-                       indices_trial, basis_trial,
-                       basis_test, cords, test_function,
-                       is_linear,
-                       mapping):
-
-    cls = (_partial_derivatives,
-           VectorTestFunction,
-           ScalarTestFunction,
-           IndexedTestTrial)
-
-    dim  = len(indices_test)
-
-    if not isinstance(atom, cls):
-        raise TypeError('atom must be of type {}'.format(str(cls)))
-
-    orders = [0 for i in range(0, dim)]
-    p_indices = get_index_derivatives(atom)
-
-    # ...
-    def _get_name(atom):
-        atom_name = None
-        if isinstance( atom, ScalarTestFunction ):
-            atom_name = str(atom.name)
-
-        elif isinstance( atom, VectorTestFunction ):
-            atom_name = str(atom.name)
-
-        elif isinstance( atom, IndexedTestTrial ):
-            atom_name = str(atom.base.name)
-
-        else:
-            raise TypeError('> Wrong type')
-
-        return atom_name
-    # ...
-
-    if isinstance(atom, _partial_derivatives):
-        a = get_atom_derivatives(atom)
-        atom_name = _get_name(a)
-
-        orders[atom.grad_index] = p_indices[atom.coordinate]
-
-    else:
-        atom_name = _get_name(atom)
-
-    test_names = [_get_name(i) for i in test_function]
-    test = atom_name in test_names
-
-    if test or is_linear:
-        basis  = basis_test
-        idxs   = indices_test
-    else:
-        basis  = basis_trial
-        idxs   = indices_trial
-
-    args = []
-    for i in range(dim):
-        args.append(basis[i][idxs[i],orders[i],indices_quad[i]])
-
-    # ... assign basis on quad point
-    logical = not( mapping is None )
-    name = print_expression(atom, logical=logical)
-    assign = Assign(Symbol(name), Mul(*args))
-    # ...
-
-    # ... map basis function
-    map_stmts = []
-    if mapping and  isinstance(atom, _partial_derivatives):
-        name = print_expression(atom)
-        rhs = LogicalExpr(mapping, atom)
-        rhs = SymbolicExpr(rhs)
-        map_stmts = [Assign(Symbol(name), rhs)]
-    # ...
-
-    return assign, map_stmts
-
-
-#==============================================================================
-def compute_atoms_expr_field(atom, indices_quad,
-                            idxs, basis,
-                            test_function, mapping):
-
-    if not is_field(atom):
-        raise TypeError('atom must be a field expr')
-
-    field = list(atom.atoms(ScalarField))[0]
-    field_name = 'coeff_'+str(field.name)
-
-    # ...
-    if isinstance(atom, _partial_derivatives):
-        direction = atom.grad_index + 1
-
-    else:
-        direction = 0
-    # ...
-
-    # ...
-    test_function = atom.subs(field, test_function)
-    name = print_expression(test_function)
-    test_function = Symbol(name)
-    # ...
-
-    # ...
-    args = []
-    dim  = len(idxs)
-    for i in range(dim):
-        if direction == i+1:
-            args.append(basis[i][idxs[i],1,indices_quad[i]])
-
-        else:
-            args.append(basis[i][idxs[i],0,indices_quad[i]])
-
-    init = Assign(test_function, Mul(*args))
-    # ...
-
-    # ...
-    args = [IndexedBase(field_name)[idxs], test_function]
-
-    val_name = print_expression(atom) + '_values'
-    val  = IndexedBase(val_name)[indices_quad]
-    update = AugAssign(val,'+',Mul(*args))
-    # ...
-
-    # ... map basis function
-    map_stmts = []
-    if mapping and  isinstance(atom, _partial_derivatives):
-        name = print_expression(atom)
-        rhs = LogicalExpr(mapping, atom)
-        rhs = SymbolicExpr(rhs)
-        map_stmts = [Assign(Symbol(name), rhs)]
-    # ...
-
-    return init, update, map_stmts
-
-
-#==============================================================================
-def compute_atoms_expr_vector_field(atom, indices_quad,
-                            idxs, basis,
-                            test_function, mapping):
-
-    if not is_vector_field(atom):
-        raise TypeError('atom must be a vector field expr')
-
-
-    vector_field = atom
-    vector_field_name = 'coeff_' + print_expression(get_atom_derivatives(atom))
-
-    # ...
-    if isinstance(atom, _partial_derivatives):
-        direction = atom.grad_index + 1
-
-    else:
-        direction = 0
-    # ...
-
-    # ...
-    base = list(atom.atoms(VectorField))[0]
-    test_function = atom.subs(base, test_function)
-    name = print_expression(test_function)
-    test_function = Symbol(name)
-    # ...
-
-    # ...
-    args = []
-    dim  = len(idxs)
-    for i in range(dim):
-        if direction == i+1:
-            args.append(basis[i][idxs[i],1,indices_quad[i]])
-
-        else:
-            args.append(basis[i][idxs[i],0,indices_quad[i]])
-
-    init = Assign(test_function, Mul(*args))
-    # ...
-
-    # ...
-    args = [IndexedBase(vector_field_name)[idxs], test_function]
-    val_name = print_expression(atom) + '_values'
-    val  = IndexedBase(val_name)[indices_quad]
-    update = AugAssign(val,'+',Mul(*args))
-    # ...
-
-    # ... map basis function
-    map_stmts = []
-    if mapping and  isinstance(atom, _partial_derivatives):
-        name = print_expression(atom)
-        rhs = LogicalExpr(mapping, atom)
-        rhs = SymbolicExpr(rhs)
-        map_stmts = [Assign(Symbol(name), rhs)]
-    # ...
-
-    return init, update, map_stmts
-
-
-#==============================================================================
-def compute_atoms_expr_mapping(atom, indices_quad,
-                               idxs, basis,
-                               test_function):
-
-    _print = lambda i: print_expression(i, mapping_name=False)
-
-    element = get_atom_derivatives(atom)
-    element_name = 'coeff_' + _print(element)
-
-    # ...
-    if isinstance(atom, _partial_derivatives):
-        direction = atom.grad_index + 1
-
-    else:
-        direction = 0
-    # ...
-
-    # ...
-    test_function = atom.subs(element, test_function)
-    name = print_expression(test_function, logical=True)
-    test_function = Symbol(name)
-    # ...
-
-    # ...
-    args = []
-    dim  = len(idxs)
-    for i in range(dim):
-        if direction == i+1:
-            args.append(basis[i][idxs[i],1,indices_quad[i]])
-
-        else:
-            args.append(basis[i][idxs[i],0,indices_quad[i]])
-
-    init = Assign(test_function, Mul(*args))
-    # ...
-
-    # ...
-    args = [IndexedBase(element_name)[idxs], test_function]
-    val_name = _print(atom) + '_values'
-    val  = IndexedBase(val_name)[indices_quad]
-    update = AugAssign(val,'+',Mul(*args))
-    # ...
-
-    return init, update
-
-#==============================================================================
-def rationalize_eval_mapping(mapping, nderiv, space, indices_quad):
-
-    _print = lambda i: print_expression(i, mapping_name=False)
-
-    M = mapping
-    dim = space.ldim
-    ops = _partial_derivatives[:dim]
-
-    # ... mapping components and their derivatives
-    components = [M[i] for i in range(0, dim)]
-    elements = list(components)
-
-    if nderiv > 0:
-        elements += [d(M[i]) for d in ops for i in range(0, dim)]
-
-    if nderiv > 1:
-        elements += [d1(d2(M[i])) for e,d1 in enumerate(ops)
-                                  for d2 in ops[:e+1]
-                                  for i in range(0, dim)]
-
-    if nderiv > 2:
-        raise NotImplementedError('TODO')
-    # ...
-
-    # ... weights and their derivatives
-    # TODO check if 'w' exist already
-    weights = ScalarField(space, name='w')
-
-    weights_elements = [weights]
-    if nderiv > 0:
-        weights_elements += [d(weights) for d in ops]
-
-    if nderiv > 1:
-        weights_elements += [d1(d2(weights)) for e,d1 in enumerate(ops)
-                                             for d2 in ops[:e+1]]
-
-    if nderiv > 2:
-        raise NotImplementedError('TODO')
-    # ...
-
-    stmts = []
-    # declarations
-    stmts += [Comment('declarations')]
-    for atom in elements + weights_elements:
-        atom_name = _print(atom)
-        val_name = atom_name + '_values'
-        val  = IndexedBase(val_name)[indices_quad]
-
-        stmt = Assign(atom_name, val)
-        stmts += [stmt]
-
-    # assignements
-    stmts += [Comment('rationalize')]
-
-    # 0 order terms
-    w  = Symbol( _print( weights ) )
-    for i in range(dim):
-        u  = Symbol( _print( M[i] ) )
-
-        val_name = u.name + '_values'
-        val  = IndexedBase(val_name)[indices_quad]
-        stmt = Assign(val, u / w )
-
-        stmts += [stmt]
-
-    # 1 order terms
-    if nderiv >= 1:
-        for d in ops:
-            w  = Symbol( _print( weights    ) )
-            dw = Symbol( _print( d(weights) ) )
-
-            for i in range(dim):
-                u  = Symbol( _print( M[i]    ) )
-                du = Symbol( _print( d(M[i]) ) )
-
-                val_name = du.name + '_values'
-                val  = IndexedBase(val_name)[indices_quad]
-                stmt = Assign(val, du / w - u * dw / w**2 )
-
-                stmts += [stmt]
-
-    # 2 order terms
-    if nderiv >= 2:
-        raise NotImplementedError("")
-
-    return stmts
-
-#==============================================================================
 # TODO take exponent to 1/dim
 def area_eval_mapping(mapping, area, dim, indices_quad, weight):
 
@@ -767,631 +241,6 @@ def area_eval_mapping(mapping, area, dim, indices_quad, weight):
     return stmts
 
 
-#==============================================================================
-def is_field(expr):
-
-    if isinstance(expr, _partial_derivatives):
-        return is_field(expr.args[0])
-
-    elif isinstance(expr, ScalarField):
-        return True
-
-    return False
-
-#==============================================================================
-def is_vector_field(expr):
-
-    if isinstance(expr, _partial_derivatives):
-        return is_vector_field(expr.args[0])
-
-    elif isinstance(expr, (VectorField, IndexedVectorField)):
-        return True
-
-    return False
-
-
-#==============================================================================
-class SplBasic(Basic):
-    _discrete_boundary = None
-
-    def __new__(cls, tag, name=None, prefix=None, debug=False, detailed=False,
-                mapping=None, is_rational_mapping=None):
-
-        if name is None:
-            if prefix is None:
-                raise ValueError('prefix must be given')
-
-            name = '{prefix}_{tag}'.format(tag=tag, prefix=prefix)
-
-        obj = Basic.__new__(cls)
-
-        obj._name                = name
-        obj._tag                 = tag
-        obj._dependencies        = []
-        obj._debug               = debug
-        obj._detailed            = detailed
-        obj._mapping             = mapping
-        obj._is_rational_mapping = is_rational_mapping
-        obj._imports = []
-
-        return obj
-
-    @property
-    def name(self):
-        return self._name
-
-    @property
-    def tag(self):
-        return self._tag
-
-    @property
-    def func(self):
-        return self._func
-
-    @property
-    def basic_args(self):
-        return self._basic_args
-
-    @property
-    def dependencies(self):
-        return self._dependencies
-
-    @property
-    def debug(self):
-        return self._debug
-
-    @property
-    def detailed(self):
-        return self._detailed
-
-    @property
-    def mapping(self):
-        return self._mapping
-
-    @property
-    def is_rational_mapping(self):
-        return self._is_rational_mapping
-
-    @property
-    def discrete_boundary(self):
-        return self._discrete_boundary
-
-    @property
-    def imports(self):
-        return self._imports
-
-
-#==============================================================================
-class EvalMapping(SplBasic):
-
-    def __new__(cls, space, mapping, discrete_boundary=None, name=None,
-                boundary_basis=None, nderiv=1, is_rational_mapping=None,
-                area=None, backend=None):
-
-        if not isinstance(mapping, Mapping):
-            raise TypeError('> Expecting a Mapping object')
-
-        obj = SplBasic.__new__(cls, mapping, name=name,
-                               prefix='eval_mapping', mapping=mapping,
-                               is_rational_mapping=is_rational_mapping)
-
-        obj._space = space
-        obj._discrete_boundary = discrete_boundary
-        obj._boundary_basis = boundary_basis
-        obj._backend = backend
-
-        dim = mapping.rdim
-
-        # ...
-        lcoords = ['x1', 'x2', 'x3'][:dim]
-        obj._lcoords = symbols(lcoords)
-        # ...
-
-        # ...
-        ops = _partial_derivatives[:dim]
-        M = mapping
-
-        components = [M[i] for i in range(0, dim)]
-
-        d_elements = {}
-        d_elements[0] = list(components)
-
-        if nderiv > 0:
-            ls = [d(M[i]) for d in ops for i in range(0, dim)]
-
-            d_elements[1] = ls
-
-        if nderiv > 1:
-            ls = [d1(d2(M[i])) for e,d1 in enumerate(ops)
-                               for d2 in ops[:e+1]
-                               for i in range(0, dim)]
-
-            d_elements[2] = ls
-
-        if nderiv > 2:
-            raise NotImplementedError('TODO')
-
-        elements = [i for l in d_elements.values() for i in l]
-        obj._elements = tuple(elements)
-        obj._d_elements = d_elements
-
-        obj._components = tuple(components)
-        obj._nderiv = nderiv
-        obj._area = area
-        # ...
-
-        obj._func = obj._initialize()
-
-        return obj
-
-    @property
-    def space(self):
-        return self._space
-
-    @property
-    def boundary_basis(self):
-        return self._boundary_basis
-
-    @property
-    def nderiv(self):
-        return self._nderiv
-
-    @property
-    def lcoords(self):
-        return self._lcoords
-
-    @property
-    def elements(self):
-        return self._elements
-
-    @property
-    def d_elements(self):
-        return self._d_elements
-
-    @property
-    def components(self):
-        return self._components
-
-    @property
-    def mapping_coeffs(self):
-        return self._mapping_coeffs
-
-    @property
-    def mapping_values(self):
-        return self._mapping_values
-
-    @property
-    def backend(self):
-        return self._backend
-
-    @property
-    def area(self):
-        return self._area
-
-    @property
-    def weights(self):
-        return self._weights
-
-    def build_arguments(self, data):
-
-        other = data
-        if self.area:
-            other = other + self.weights + (self.area, )
-
-        return self.basic_args + other
-
-    def _initialize(self):
-        space = self.space
-        dim = space.ldim
-
-        _print = lambda i: print_expression(i, mapping_name=False)
-        mapping_atoms = [_print(f) for f in self.components]
-        mapping_str = [_print(f) for f in self.elements]
-#        mapping_str = sorted([_print(f) for f in self.elements])
-
-        # ... declarations
-        degrees        = variables( 'p1:%s'%(dim+1), 'int')
-        orders         = variables( 'k1:%s'%(dim+1), 'int')
-        indices_basis  = variables( 'jl1:%s'%(dim+1), 'int')
-        indices_quad   = variables( 'g1:%s'%(dim+1), 'int')
-        basis          = variables('basis1:%s'%(dim+1),
-                                  dtype='real', rank=3, cls=IndexedVariable)
-        mapping_coeffs = variables(['coeff_{}'.format(f) for f in mapping_atoms],
-                                  dtype='real', rank=dim, cls=IndexedVariable)
-        mapping_values = variables(['{}_values'.format(f) for f in mapping_str],
-                                  dtype='real', rank=dim, cls=IndexedVariable)
-
-        # ... needed for area
-        weights = variables('quad_w1:%s'%(dim+1),
-                            dtype='real', rank=1, cls=IndexedVariable)
-
-        self._weights = weights
-        # ...
-
-        weights_elements = []
-        if self.is_rational_mapping:
-            # TODO check if 'w' exist already
-            weights_pts = ScalarField(self.space, name='w')
-
-            weights_elements = [weights_pts]
-
-            # ...
-            nderiv = self.nderiv
-            ops = _partial_derivatives[:dim]
-
-            if nderiv > 0:
-                weights_elements += [d(weights_pts) for d in ops]
-
-            if nderiv > 1:
-                weights_elements += [d1(d2(weights_pts)) for e,d1 in enumerate(ops)
-                                                     for d2 in ops[:e+1]]
-
-            if nderiv > 2:
-                raise NotImplementedError('TODO')
-            # ...
-
-            mapping_weights_str = [_print(f) for f in weights_elements]
-            mapping_wvalues = variables(['{}_values'.format(f) for f in mapping_weights_str],
-                                                dtype='real', rank=dim, cls=IndexedVariable)
-
-            mapping_coeffs  = mapping_coeffs + (IndexedVariable('coeff_w', dtype='real', rank=dim),)
-            mapping_values  = mapping_values + tuple(mapping_wvalues)
-
-        weights_elements = tuple(weights_elements)
-        # ...
-
-        # ... ranges
-        ranges_basis = [Range(degrees[i]+1) for i in range(dim)]
-        ranges_quad  = [Range(orders[i]) for i in range(dim)]
-        # ...
-
-        # ... basic arguments
-        self._basic_args = (orders)
-        # ...
-
-        # ...
-        self._mapping_coeffs  = mapping_coeffs
-        self._mapping_values  = mapping_values
-        # ...
-
-        # ...
-        Nj = ScalarTestFunction(space, name='Nj')
-        body = []
-        init_basis = OrderedDict()
-        updates = []
-        for atom in self.elements + weights_elements:
-            init, update = compute_atoms_expr_mapping(atom, indices_quad,
-                                                      indices_basis, basis, Nj)
-
-            updates.append(update)
-
-            basis_name = str(init.lhs)
-            init_basis[basis_name] = init
-
-        init_basis = OrderedDict(sorted(init_basis.items()))
-        body += list(init_basis.values())
-        body += updates
-        # ...
-
-        # put the body in tests for loops
-        body = filter_loops(indices_basis, ranges_basis, body,
-                            self.discrete_boundary,
-                            boundary_basis=self.boundary_basis)
-
-        if self.is_rational_mapping:
-            stmts = rationalize_eval_mapping(self.mapping, self.nderiv,
-                                             self.space, indices_quad)
-
-            body += stmts
-
-        # ...
-        if self.area:
-            weight = filter_product(indices_quad, weights, self.discrete_boundary)
-
-            stmts = area_eval_mapping(self.mapping, self.area, dim, indices_quad, weight)
-
-            body += stmts
-        # ...
-
-        # put the body in for loops of quadrature points
-        body = filter_loops(indices_quad, ranges_quad, body,
-                            self.discrete_boundary,
-                            boundary_basis=self.boundary_basis)
-
-        # initialization of the matrix
-        init_vals = [f[[Slice(None,None)]*dim] for f in mapping_values]
-        init_vals = [Assign(e, 0.0) for e in init_vals]
-        body =  init_vals + body
-
-        if self.area:
-            # add init to 0 at the begining
-            body = [Assign(self.area, 0.0)] + body
-
-            # add power to 1/dim
-            body += [Assign(self.area, Pow(self.area, 1./dim))]
-
-        func_args = self.build_arguments(degrees + basis + mapping_coeffs + mapping_values)
-
-        decorators = {}
-        header = None
-        if self.backend['name'] == 'pyccel':
-            decorators = {'types': build_types_decorator(func_args)}
-        elif self.backend['name'] == 'numba':
-            decorators = {'jit':[]}
-        elif self.backend['name'] == 'pythran':
-            header = build_pythran_types_header(self.name, func_args)
-
-        return FunctionDef(self.name, list(func_args), [], body,
-                           decorators=decorators,header=header)
-
-
-#==============================================================================
-class EvalField(SplBasic):
-
-    def __new__(cls, space, fields, discrete_boundary=None, name=None,
-                boundary_basis=None, mapping=None, is_rational_mapping=None,backend=None):
-
-        if not isinstance(fields, (tuple, list, Tuple)):
-            raise TypeError('> Expecting an iterable')
-
-        obj = SplBasic.__new__(cls, space, name=name,
-                               prefix='eval_field', mapping=mapping,
-                               is_rational_mapping=is_rational_mapping)
-
-        obj._space = space
-        obj._fields = Tuple(*fields)
-        obj._discrete_boundary = discrete_boundary
-        obj._boundary_basis = boundary_basis
-        obj._backend = backend
-        obj._func = obj._initialize()
-
-
-        return obj
-
-    @property
-    def space(self):
-        return self._space
-
-    @property
-    def fields(self):
-        return self._fields
-
-    @property
-    def map_stmts(self):
-        return self._map_stmts
-
-    @property
-    def boundary_basis(self):
-        return self._boundary_basis
-
-    @property
-    def backend(self):
-        return self._backend
-
-    def build_arguments(self, data):
-
-        other = data
-
-        return self.basic_args + other
-
-    def _initialize(self):
-        space = self.space
-        dim = space.ldim
-        mapping = self.mapping
-
-        field_atoms = self.fields.atoms(ScalarField)
-        fields_str = sorted([print_expression(f) for f in self.fields])
-
-        # ... declarations
-        degrees       = variables( 'p1:%s'%(dim+1), 'int')
-        orders        = variables( 'k1:%s'%(dim+1), 'int')
-        indices_basis = variables( 'jl1:%s'%(dim+1), 'int')
-        indices_quad  = variables( 'g1:%s'%(dim+1), 'int')
-        basis         = variables('basis1:%s'%(dim+1),
-                                  dtype='real', rank=3, cls=IndexedVariable)
-        fields_coeffs = variables(['coeff_{}'.format(f) for f in field_atoms],
-                                  dtype='real', rank=dim, cls=IndexedVariable)
-        fields_val    = variables(['{}_values'.format(f) for f in fields_str],
-                                  dtype='real', rank=dim, cls=IndexedVariable)
-        # ...
-
-        # ... ranges
-        ranges_basis = [Range(degrees[i]+1) for i in range(dim)]
-        ranges_quad  = [Range(orders[i]) for i in range(dim)]
-        # ...
-
-        # ... basic arguments
-        self._basic_args = (orders)
-        # ...
-
-        # ...
-        Nj = ScalarTestFunction(space, name='Nj')
-        body = []
-        init_basis = OrderedDict()
-        init_map   = OrderedDict()
-        updates = []
-        for atom in self.fields:
-            init, update, map_stmts = compute_atoms_expr_field(atom, indices_quad, indices_basis,
-                                                               basis, Nj,
-                                                               mapping=mapping)
-
-            updates.append(update)
-
-            basis_name = str(init.lhs)
-            init_basis[basis_name] = init
-            for stmt in map_stmts:
-                init_map[str(stmt.lhs)] = stmt
-
-        init_basis = OrderedDict(sorted(init_basis.items()))
-        body += list(init_basis.values())
-        body += updates
-        self._map_stmts = init_map
-        # ...
-
-        # put the body in tests for loops
-        body = filter_loops(indices_basis, ranges_basis, body,
-                            self.discrete_boundary,
-                            boundary_basis=self.boundary_basis)
-
-
-        # put the body in for loops of quadrature points
-        body = filter_loops(indices_quad, ranges_quad, body,
-                            self.discrete_boundary,
-                            boundary_basis=self.boundary_basis)
-
-
-        # initialization of the matrix
-        init_vals = [f[[Slice(None,None)]*dim] for f in fields_val]
-        init_vals = [Assign(e, 0.0) for e in init_vals]
-        body =  init_vals + body
-
-        func_args = self.build_arguments(degrees + basis + fields_coeffs + fields_val)
-
-        decorators = {}
-        header = None
-        if self.backend['name'] == 'pyccel':
-            decorators = {'types': build_types_decorator(func_args)}
-        elif self.backend['name'] == 'numba':
-            decorators = {'jit':[]}
-        elif self.backend['name'] == 'pythran':
-            header = build_pythran_types_header(self.name, func_args)
-
-        return FunctionDef(self.name, list(func_args), [], body,
-                           decorators=decorators,header=header)
-
-
-
-#==============================================================================
-class EvalVectorField(SplBasic):
-
-    def __new__(cls, space, vector_fields, discrete_boundary=None, name=None,
-                boundary_basis=None, mapping=None, is_rational_mapping=None, backend = None):
-
-        if not isinstance(vector_fields, (tuple, list, Tuple)):
-            raise TypeError('> Expecting an iterable')
-
-        obj = SplBasic.__new__(cls, space, name=name,
-                               prefix='eval_vector_field', mapping=mapping,
-                               is_rational_mapping=is_rational_mapping)
-
-        obj._space = space
-        obj._vector_fields = Tuple(*vector_fields)
-        obj._discrete_boundary = discrete_boundary
-        obj._boundary_basis = boundary_basis
-        obj._backend = backend
-        obj._func = obj._initialize()
-
-        return obj
-
-    @property
-    def space(self):
-        return self._space
-
-    @property
-    def vector_fields(self):
-        return self._vector_fields
-
-    @property
-    def map_stmts(self):
-        return self._map_stmts
-
-    @property
-    def boundary_basis(self):
-        return self._boundary_basis
-
-    @property
-    def backend(self):
-        return self._backend
-
-    def build_arguments(self, data):
-
-        other = data
-
-        return self.basic_args + other
-
-    def _initialize(self):
-        space = self.space
-        dim = space.ldim
-        mapping = self.mapping
-
-        vector_field_atoms = self.vector_fields.atoms(VectorField)
-        vector_field_atoms = [f[i] for f in vector_field_atoms for i in range(0, dim)]
-        vector_fields_str = sorted([print_expression(f) for f in self.vector_fields])
-
-        # ... declarations
-        degrees       = variables('p1:%s'%(dim+1),  'int')
-        orders        = variables('k1:%s'%(dim+1),  'int')
-        indices_basis = variables('jl1:%s'%(dim+1), 'int')
-        indices_quad  = variables('g1:%s'%(dim+1),  'int')
-        basis         = variables('basis1:%s'%(dim+1),
-                                  dtype='real', rank=3, cls=IndexedVariable)
-        coeffs = ['coeff_{}'.format(print_expression(f)) for f in vector_field_atoms]
-        vector_fields_coeffs = variables(coeffs, dtype='real', rank=dim, cls=IndexedVariable)
-        vector_fields_val    = variables(['{}_values'.format(f) for f in vector_fields_str],
-                                          dtype='real', rank=dim, cls=IndexedVariable)
-        # ...
-
-        # ... ranges
-        ranges_basis = [Range(degrees[i]+1) for i in range(dim)]
-        ranges_quad  = [Range(orders[i]) for i in range(dim)]
-        # ...
-
-        # ... basic arguments
-        self._basic_args = (orders)
-        # ...
-
-        # ...
-        Nj = VectorField(space, name='Nj')
-        body = []
-        init_basis = OrderedDict()
-        init_map   = OrderedDict()
-        updates = []
-        for atom in self.vector_fields:
-            init, update, map_stmts = compute_atoms_expr_vector_field(atom, indices_quad, indices_basis,
-                                                                      basis, Nj,
-                                                                      mapping=mapping)
-
-            updates.append(update)
-
-            basis_name = str(init.lhs)
-            init_basis[basis_name] = init
-            for stmt in map_stmts:
-                init_map[str(stmt.lhs)] = stmt
-
-        init_basis = OrderedDict(sorted(init_basis.items()))
-        body += list(init_basis.values())
-        body += updates
-        self._map_stmts = init_map
-        # ...
-
-        # put the body in tests for loops
-        body = filter_loops(indices_basis, ranges_basis, body,
-                            self.discrete_boundary,
-                            boundary_basis=self.boundary_basis)
-
-        # put the body in for loops of quadrature points
-        body = filter_loops(indices_quad, ranges_quad, body,
-                            self.discrete_boundary,
-                            boundary_basis=self.boundary_basis)
-
-        # initialization of the matrix
-        init_vals = [f[[Slice(None,None)]*dim] for f in vector_fields_val]
-        init_vals = [Assign(e, 0.0) for e in init_vals]
-        body =  init_vals + body
-
-        func_args = self.build_arguments(degrees + basis + vector_fields_coeffs + vector_fields_val)
-
-        decorators = {}
-        header = None
-        if self.backend['name'] == 'pyccel':
-            decorators = {'types': build_types_decorator(func_args)}
-        elif self.backend['name'] == 'numba':
-            decorators = {'jit':[]}
-        elif self.backend['name'] == 'pythran':
-            header = build_pythran_types_header(self.name, func_args)
-
-        return FunctionDef(self.name, list(func_args), [], body,
-                           decorators=decorators,header=header)
 
 #==============================================================================
 # target is used when there are multiple expression (domain/boundaries)
@@ -3248,5 +2097,3 @@ class VectorDot(SplBasic):
 
         return FunctionDef(self.name, list(func_args), [], body,
                            decorators=decorators,header=header)
-
-
